\name{TimeAdder-package}
\alias{TimeAdder-package}
\alias{TimeAdder}
\docType{package}
\title{
\packageTitle{TimeAdder}
}
\description{
\packageDescription{TimeAdder}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{TimeAdder}
\packageIndices{TimeAdder}
~~ An overview of how to use the package, including the ~~
~~ most important functions ~~
}
\author{
\packageAuthor{TimeAdder}

Maintainer: \packageMaintainer{TimeAdder}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from ~~
~~ file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
